# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 4.0.0
  description: TIAMO BURGER — API del Sistema POS
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /mesas:
    get:
      operationId: getMesas
      tags: [mesas]
      summary: Listar todas las mesas
      responses:
        "200":
          description: Lista de mesas
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Mesa"
    post:
      operationId: crearMesa
      tags: [mesas]
      summary: Crear nueva mesa (solo admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CrearMesaInput"
      responses:
        "201":
          description: Mesa creada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mesa"
  /mesas/{numero}:
    patch:
      operationId: actualizarMesa
      tags: [mesas]
      summary: Actualizar estado de una mesa
      parameters:
        - name: numero
          in: path
          required: true
          schema:
            type: string
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActualizarMesaInput"
      responses:
        "200":
          description: Mesa actualizada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Mesa"
        "404":
          description: Mesa no encontrada
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
    delete:
      operationId: eliminarMesa
      tags: [mesas]
      summary: Eliminar mesa (solo admin)
      parameters:
        - name: numero
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Mesa eliminada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
...
  /impresora/config:
    get:
      operationId: getImpresoraConfig
      tags: [impresora]
      summary: Obtener configuración de impresora
      responses:
        "200":
          description: Config actual
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpresoraConfig"
    post:
      operationId: setImpresoraConfig
      tags: [impresora]
      summary: Actualizar configuración de impresora
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImpresoraConfig"
      responses:
        "200":
          description: Config actualizada
          content:
            application/json:
              schema:
                type: object
                properties:
                  ok:
                    type: boolean
                required: [ok]
  /impresora/imprimir:
    post:
      operationId: imprimirTicket
      tags: [impresora]
      summary: Imprimir ticket
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ImprimirInput"
      responses:
        "200":
          description: Resultado de impresión
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ImpresionResultado"
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    EstadoMesa:
      type: string
      enum: [libre, ocupada, proceso]
...
    Mesa:
      type: object
      properties:
        numero:
          type: string
        estado:
          $ref: "#/components/schemas/EstadoMesa"
        personas:
          type: integer
        nombre:
          type: string
        zona:
          type: string
      required: [numero, estado, personas]
    CrearMesaInput:
      type: object
      properties:
        numero:
          type: string
        nombre:
          type: string
        zona:
          type: string
        capacidad:
          type: integer
      required: [numero]
    ActualizarMesaInput:
      type: object
      properties:
        estado:
          $ref: "#/components/schemas/EstadoMesa"
        personas:
          type: integer
        nombre:
          type: string
        zona:
          type: string
...
    ImpresoraConfig:
      type: object
      properties:
        modo:
          type: string
          enum: [simulado, usb, wifi, ethernet, bluetooth]
        ip:
          type: string
        puerto:
          type: integer
        ssid:
          type: string
        bluetoothName:
          type: string
      required: [modo]
    ImprimirInput:
      type: object
      properties:
        ticket:
          type: string
        tipo:
          type: string
      required: [ticket]
    ImpresionResultado:
      type: object
      properties:
        ok:
          type: boolean
        modo:
          type: string
        mensaje:
          type: string
        ticket:
          type: string
        error:
          type: string
      required: [ok]
...